Bring up-to-date with upstream/master.

+++ sulong/projects/com.oracle.truffle.llvm.runtime/src/com/oracle/truffle/llvm/runtime/debug/value/LLVMSourceTypeFactory.java
@@ -220,7 +220,7 @@ public final class LLVMSourceTypeFactory {
 
         private long getBitSize(Type type) {
             final int byteSize = context.getByteSize(type);
-            return byteSize * Byte.SIZE;
+            return byteSize * (long) Byte.SIZE;
         }
 
         private long getAlignment(Type type) {
